+++ CHANGELOG.md
@@ -1,4 +1,18 @@
-﻿## Version 4.02 [05-02-2026]
+﻿## Version 4.03 [11-02-2026]
+
+**All WoW Versions:**  
+- ElvUI 15.05+ compatibility  
+- Added new option for instance-only nameplate classification coloring in the ElvUI Tweaks config tab  
+- Fixed a bug which caused color coding to not be included when copy-pasting text formats  
+- Updated deDE locales (Credits: Dlarge)  
+- Updated the default value for the CVar nameplateOverlapV (from 2.2 to 1.6)  
+
+**Retail:**  
+- Initial default filters based on new patch 12.0.1 APIs (Will tweak further if needed)  
+- Updated nameplate CVars to class color friendly name-only plates in instances  
+- Updated TOC for patch 12.0.1  
+
+## Version 4.02 [05-02-2026]
 
 **All WoW Versions:**  
 - ElvUI 15.03+ compatibility